<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #46579

Problem Summary: planner: optimize the performance of `PlanCacheParamList.String()`

### What is changed and how it works?

planner: optimize the performance of `PlanCacheParamList.String()`

```
Before
  BenchmarkDatumsToString-8                 676939              1605 ns/op             952 B/op         37 allocs/op
  BenchmarkDatumsToStringStr-8              315628              3564 ns/op            1184 B/op          4 allocs/op
  BenchmarkDatumsToStringLongStr-8           17454             68606 ns/op           21794 B/op          4 allocs/op

After
  BenchmarkDatumsToString-8                1000000              1044 ns/op             792 B/op         34 allocs/op
  BenchmarkDatumsToStringStr-8            10546382               113.5 ns/op           584 B/op          2 allocs/op
  BenchmarkDatumsToStringLongStr-8         3918226               300.8 ns/op          2320 B/op          3 allocs/op

```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```