@dbsid 2023-09-01
after the fix, in the tidb cpu profile the, no significant usages from PlanCacheParamList

![image](https://github.com/pingcap/tidb/assets/2316425/92bedbf9-a311-49cc-bcac-47e9c547f292)